requirements.txt
playwright
flask
pandas
gunicorn
schedule
python-dotenv
flask-sqlalchemy
numpy
